ml/requirements.txt
# Core
numpy>=1.23
pandas>=1.5
scipy>=1.9

# Imaging
opencv-python>=4.7
Pillow>=9.5
scikit-image>=0.21

# ML
torch>=2.0
torchvision>=0.15
scikit-learn>=1.2

# Utils
tqdm
pyyaml
matplotlib
python-dotenv

